requirements.txt
streamlit>=1.41.0
openai>=1.60.0
pypdf>=5.1.0
PyYAML>=6.0.2
